requirements.txt
astroid==2.5.2
autopep8==1.5.6
certifi==2020.12.5
chardet==4.0.0
geographiclib==1.50
geopy==2.1.0
idna==2.10
isort==5.8.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
numpy==1.20.2
pycodestyle==2.7.0
pylint==2.7.4
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
timezonefinder==5.2.0
toml==0.10.2
urllib3==1.26.4
wrapt==1.12.1
